% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stepwise.R
\docType{data}
\name{Compare}
\alias{Compare}
\title{Comparison decomposition results by age and cause}
\format{A matrix with 19 rows and 6 columns}
\source{
\url{https://www.demogr.mpg.de/en/projects_publications/publications_1904/mpidr_technical_reports/an_excel_spreadsheet_for_the_decomposition_of_a_difference_between_two_values_of_an_aggregate_4591.htm}
}
\usage{
Compare
}
\description{
A matrix containing the contributions to the difference in life expectancy at birth between 2002 US males and England and Wales males. Ages (in rows) are in abridged categories, 0-85, and there are six causes, including other, in columns. The sum of the matrix is the difference in life expectancy at birth between the two populations. Values are based on symmetrical stepwise replacement from young to old ages only. This is just to make sure implementation is close.
}
\keyword{datasets}
