% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stepwise.R
\docType{data}
\name{Mxc2}
\alias{Mxc2}
\title{Year 2002 death rates by cause for England and Wales males in abridged age classes}
\format{A matrix with 19 rows and 6 columns}
\source{
\url{https://www.demogr.mpg.de/en/projects_publications/publications_1904/mpidr_technical_reports/an_excel_spreadsheet_for_the_decomposition_of_a_difference_between_two_values_of_an_aggregate_4591.htm}
}
\usage{
Mxc2
}
\description{
A matrix containing death rates for six causes (one of which is other) for abrdged age classes 0-85. Ages are labelled in rows, and causes in column names.
}
\keyword{datasets}
