% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stepwise.R
\name{Mxc2e0abr}
\alias{Mxc2e0abr}
\title{get life expectancy at birth from an (abridged)age-cause matrix}
\usage{
Mxc2e0abr(Mxc)
}
\arguments{
\item{Mxc}{numeric matrix}
}
\value{
numeric life expectancy at birth
}
\description{
Given a matrix with abridged ages in rows and causes of death in columns, then calculate life expectancy at birth using \code{LTabr()}.
}
\details{
This assumes that the marginal row sums give all-cause mortality rates. Give an other category if you need to top-up to all-cause mortality. Do not include all-cause mortality itself!
}
