% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Horiuchi.R
\docType{data}
\name{rates2}
\alias{rates2}
\title{Fake data generated for example.}
\format{An object of class \code{matrix} with 101 rows and 2 columns.}
\usage{
rates2
}
\description{
The first column \code{Lx} is a discrete survival function for time point 2. The second column \code{Fx} are age specific fertility rates.
}
\examples{
\dontrun{
data(rates1)
data(rates2)
# nothing fancy
# compare Lx
plot(rates1[,1],type='l',col="blue")
lines(rates2[,1],col="green")
# compare Fx
plot(rates1[,2],type='l',col="blue")
lines(rates2[,2],col="green") 
}

}
\keyword{datasets}
